patch(robo/cli): build public directory even without vite
